+++ src/Facades/Laranail.php
@@ -22,8 +22,6 @@ use Simtabi\Laranail\Toolkit\ToolkitManager;
  * @method static \Simtabi\Laranail\Toolkit\Services\Contracts\RouteServiceInterface                 route()
  * @method static \Simtabi\Laranail\Toolkit\Services\Contracts\ValidationServiceInterface            validation()
  * @method static \Simtabi\Laranail\Toolkit\Services\Contracts\SessionServiceInterface               session()
- * @method static \Simtabi\Laranail\Toolkit\Services\Contracts\DatabaseServiceInterface              db()
- * @method static \Simtabi\Laranail\Toolkit\Services\Contracts\DatabaseServiceInterface              database()
  * @method static \Simtabi\Laranail\Toolkit\Services\Contracts\FileServiceInterface                  file()
  * @method static \Simtabi\Laranail\Toolkit\Services\Contracts\SystemServiceInterface                system()
  * @method static \Simtabi\Laranail\Toolkit\Services\ModelService                                    model()